Moved ActionListener for UI Timer into its own class Made method addListeners to add all Listeners to their corresponding elements
